Clean up fastq-files with trimmomatic
A user provides an input directory containing forward and reverse fastq-files for 1 or multiple samples. For each fastq-pair (forward+reverse) found, Illumina RNA-adapters are clipped, very low quality leading and trailing bases are trimmed and a sliding window trimming is applied on the trailing bases. Only reads surviving trimmomatic procedure with at last 36 base pairs in the read are kept. Both read pairs need to survive the trimmomatic cleaning in order to be saved.